fix dialogs in vesting page closes when data is revalidated in background
