+++ components/Vesting/types.ts
@@ -21,3 +21,12 @@ export interface IVestingGnosisFormValues {
     startDate: string;
   }[];
 }
+
+export interface IChartValues {
+  tokenSymbol: string;
+  amount: number;
+  vestingPeriod: number;
+  cliffPeriod: number | null;
+  startTime: Date;
+  vestedDays?: string | null | undefined;
+}
